# HTTPServer/http-api-spec.yaml
openapi: 3.0.0
info:
  title: "Game Logic Gateway API"
  description: "클라이언트와 HTTPS 서버를 위한 REST API 명세서"
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: "로컬 개발 환경"
paths:
...
  /api/signup:
    post:
      summary: "신규 계정 생성"
      tags:
        - Auth
      description: "ID와 비밀번호를 받아 새로운 유저 계정을 생성합니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: "계정 생성 성공 (자동 로그인 처리되어 세션 키 발급)"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthData'
        '409':
          description: "이미 존재하는 ID (Conflict)"

  # 로그인 API
  /api/login:
    post:
      summary: "로그인 요청"
      tags:
        - Auth
      description: "ID와 비밀번호로 로그인을 시도합니다. 기존 세션이 있으면 파기하고 새 세션을 발급하지만, 진행 중인 게임이 있으면 거부합니다. 매칭 대기 중이던 티켓은 자동 취소됩니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: "로그인 성공"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthData'
        '401':
          description: "인증 실패 (비밀번호 불일치 또는 존재하지 않는 ID)"
        '409':
          description: "진행 중인 게임이 있거나 매칭이 성사되어 로그인할 수 없음 (ERR_ALREADY_IN_GAME)"

  # 게스트 로그인 API [추가]
...
  /api/logout:
    post:
      summary: "로그아웃 요청"
      tags:
        - Auth
      security:
        - SessionAuth: []
      description: "현재 접속 중인 세션을 파기하고 로그아웃합니다."
      responses:
        '200':
          description: "로그아웃 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'

  # 세션 유지 (매치 종료 후 로비 복귀) [추가]
  /api/session/resume:
    post:
      summary: "세션 유지 확인 및 로비 데이터 재조회"
      tags:
        - Auth
      security:
        - SessionAuth: []
      description: |
        매치 종료 후 로비 복귀 시 기존 세션이 유효한지 확인하고, 매치 결과가 반영된
        최신 재화/인벤토리를 반환합니다 (클라이언트의 Login 과정 스킵용).
        세션 TTL은 인증 미들웨어가 슬라이딩 갱신합니다.
        shopItems는 재전송하지 않습니다 (로그인 시 받은 클라이언트 캐시 유지).
      responses:
        '200':
          description: "세션 유효 — 로비 데이터 반환"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/SessionResumeData'
        '401':
          description: "세션 없음 또는 만료 — 클라이언트는 일반 로그인 플로우로 폴백"

  # 매치메이킹 시작
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: "로그인 시 발급받은 세션 키를 입력해주세요."
  schemas:
...
    CommonResponse:
      type: object
      required: [success, code]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: integer
          example: 200
        data:
          type: object
          nullable: true
          description: "API 호출 결과 데이터 (API마다 구조가 다름)"
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      nullable: true
      properties:
        message: { type: string, example: "Invalid request" }
        code: { type: string, example: "ERR_BAD_REQUEST" }

    # ----------------------------------------------------
    # 공통 응답에 담을 진짜 내용물
    # ----------------------------------------------------
...
    AuthData:
      type: object
      properties:
        sessionId:
          type: string
          description: "API 인증에 사용할 세션 키"
          example: "sess_abc123xyz456"
        uid: # [추가] 실제 코드에서 반환하는 uid 추가
          type: integer
          description: "데이터베이스 고유 식별자 (PK)"
          example: 13
        money:
          type: integer
          description: "유저 보유 재화"
          example: 0
        inventory:
          type: array
          description: "유저 보유 아이템 목록"
          items:
            $ref: '#/components/schemas/InventoryItem'
        shopItems:
          type: array
          nullable: true
          description: "구매 가능한 아이템 목록 (게스트 로그인 시 미포함)"
          items:
            $ref: '#/components/schemas/ShopItem'
    SessionResumeData: # [추가] 세션 유지 응답 (매치 종료 후 로비 복귀)
      type: object
      properties:
        uid:
          type: integer
          description: "데이터베이스 고유 식별자 (PK)"
          example: 13
        money:
          type: integer
          description: "유저 보유 재화 (매치 결과 반영 후)"
          example: 1500
        inventory:
          type: array
          description: "매치 결과가 반영된 최신 인벤토리 (게스트는 빈 배열)"
          items:
            $ref: '#/components/schemas/InventoryItem'
    ShopItem:
      type: object
      properties:
        item_id:
          type: integer
          description: "아이템 고유 ID"
          example: 1
        price:
          type: integer
          description: "아이템 구매 가격"
          example: 500
    InventoryItem:
      type: object
      properties:
        item_id:
          type: integer
          description: "아이템 고유 ID"
          example: 101
        slot_index:
          type: integer
          description: "인벤토리 슬롯 번호"
          example: 0
        quantity:
          type: integer
          description: "보유 수량"
          example: 1
...
    AuthRequest:
      type: object
      required:
        - id
        - password
      properties:
        id:
          type: string
          description: "아이디. alphanumeric 4자 이상, 16자 이하"
          example: "tetepiti149"
          minLength: 4
          maxLength: 16
          pattern: "^[a-zA-Z0-9]+$"
        password:
          type: string
          format: password
          description: "유저 비밀번호. 서버에서 해싱 처리 예정."
          example: "q1w2e3r4!@"
          minLength: 4
          maxLength: 16
...